Apply code review changes - refactor species widget
